requirements.txt
asgiref==3.5.2
Django~=4.1.0
Pillow==9.2.0
sqlparse==0.4.2
django-braces==1.15.0
django-embed-video==1.4.4
pymemcache==3.5.2
django-debug-toolbar==3.6.0
redis==4.3.4
django-redisboard==8.3.0
djangorestframework==3.13.1
requests==2.28.1
channels==3.0.5
channels-redis==3.4.1
psycopg2==2.9.3
uwsgi==2.0.21
daphne==3.0.2
django-cors-headers==3.10.0
gunicorn==20.1.0
django-allauth==0.45.0
dj-rest-auth==2.1.11